#960 Resolve class name conflicts
