--- inventory/src/it/java/com/yas/inventory/service/LocationServiceITTemp.java
@@ -1,43 +1,0 @@
-package com.yas.inventory.service;
-
-import static com.yas.inventory.constant.TestConstants.CIRCUIT_BREAKER_NAME;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-import com.yas.inventory.viewmodel.address.AddressPostVm;
-import io.github.resilience4j.circuitbreaker.CallNotPermittedException;
-import io.github.resilience4j.circuitbreaker.CircuitBreakerRegistry;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-@SpringBootTest
-@Testcontainers
-class LocationServiceITTemp {
-    @Autowired
-    private LocationService locationService;
-    @Autowired
-    private CircuitBreakerRegistry circuitBreakerRegistry;
-
-    @Container
-    @ServiceConnection
-    static PostgreSQLContainer<?> postgresContainer = new PostgreSQLContainer<>("postgres:16");
-
-    @Test
-    void test_getAddressById_shouldThrowCallNotPermittedException_whenCircuitBreakerIsOpen() {
-        circuitBreakerRegistry.circuitBreaker(CIRCUIT_BREAKER_NAME).transitionToOpenState();
-        assertThrows(CallNotPermittedException.class, () -> locationService.getAddressById(1L));
-    }
-
-    @Test
-    void test_createAddress_shouldThrowCallNotPermittedException_whenCircuitBreakerIsOpen() {
-        AddressPostVm addressPostVm =
-            new AddressPostVm("my contact", "phone number", "address line 1", "address line 2", "city", "zip code", 1L,
-                1L, 1L);
-        circuitBreakerRegistry.circuitBreaker(CIRCUIT_BREAKER_NAME).transitionToOpenState();
-        assertThrows(CallNotPermittedException.class, () -> locationService.createAddress(addressPostVm));
-    }
-}
